+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.0-nullsafety.0]
+- Migrate to null-safety
+- When gradient is active ContinuousRectangleBorder and BeveledRectangleBorder are not supported
+- add Button style
+- Breaking: remove borderRadius
+
 ## [0.0.2] - 2021-01-18
 
 * Reformat according to dartfmt